User item fix
- MarbleBotModule takes an new JsonConstructor that takes the items dictionary as a nullable type (to avoid a null reference exception being thrown when Newtonsoft.Json tries to set MarbleBotUser.Items to null)
- Removed all summary tags
- Replaced all ushorts, shorts and uint have been with ints

+++ src/Services/DailyTimeoutService.cs
@@ -2,6 +2,6 @@
 {
     public class DailyTimeoutService
     {
-        public ushort DailyTimeout { get; set; } = 48;
+        public int DailyTimeout { get; set; } = 48;
     }
 }
